create abstract test base for sensors; refactor existing abstract sensor test.
